inventory management/warehouse/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
gunicorn>=21.2
whitenoise>=6.6
